Extract concluded event domain events to a class
So that we have one listener for this event type. Next will be moving
the integration listener to the same class for this event.
