+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/service/SNSService.kt
@@ -16,7 +16,6 @@ import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.events.ReferralEve
 import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.events.ReferralEventType
 import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.exception.AsyncEventExceptionHandling
 import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jpa.entity.Attended
-import uk.gov.justice.digital.hmpps.hmppsinterventionsservice.jpa.entity.AuthUser
 
 interface SNSService
 
@@ -80,37 +79,6 @@ class SNSReferralService(
           mapOf("referralId" to event.referral.id, "assignedTo" to (assignment.assignedTo.userName))
         )
         snsPublisher.publish(event.referral.id, assignment.assignedBy, snsEvent)
-      }
-      ReferralEventType.CANCELLED -> {
-        val snsEvent = EventDTO(
-          "intervention.referral.cancelled",
-          "A referral has been cancelled",
-          event.detailUrl,
-          event.referral.concludedAt!!,
-          mapOf("referralId" to event.referral.id)
-        )
-        snsPublisher.publish(event.referral.id, event.referral.endRequestedBy!!, snsEvent)
-      }
-      ReferralEventType.PREMATURELY_ENDED -> {
-        val snsEvent = EventDTO(
-          "intervention.referral.prematurely-ended",
-          "A referral has been ended prematurely",
-          event.detailUrl,
-          event.referral.concludedAt!!,
-          mapOf("referralId" to event.referral.id)
-        )
-        snsPublisher.publish(event.referral.id, event.referral.endRequestedBy!!, snsEvent)
-      }
-      ReferralEventType.COMPLETED -> {
-        val snsEvent = EventDTO(
-          "intervention.referral.completed",
-          "A referral has been completed",
-          event.detailUrl,
-          event.referral.concludedAt!!,
-          mapOf("referralId" to event.referral.id)
-        )
-        // This is a system generated event at present and as such the actor will represent this
-        snsPublisher.publish(event.referral.id, AuthUser.interventionsServiceUser, snsEvent)
       }
       else -> {}
     }
